Extract ResponseGatekeeper
Extract from info_request.rb and add specs
