+++ app/models/info_request.rb
@@ -1305,21 +1305,6 @@ class InfoRequest < ActiveRecord::Base
 
   private
 
-  module ResponseGatekeeper
-    class UnknownResponseGatekeeperError < ArgumentError ; end
-
-    SPECIALIZED_CLASSES = { 'nobody' => Nobody,
-                            'anybody' => Base,
-                            'authority_only' => AuthorityOnly }
-
-    def self.for(name, info_request)
-      SPECIALIZED_CLASSES.fetch(name).new(info_request)
-      rescue KeyError
-        raise UnknownResponseGatekeeperError,
-              "Unknown allow_new_responses_from '#{ name }'"
-    end
-  end
-
   module ResponseRejection
     class UnknownResponseRejectionError < ArgumentError ; end
